Split nanopub handlers from workflowhub handlers
