+++ FAIRWorkflowsExtension/nanopub_handlers.py
@@ -44,24 +44,6 @@ def nanopub_search_handler(base_url='/'):
     return endpoint, NanopubSearchHandler
 
 
-class WorkflowhubSearchHandler(APIHandler):
-
-    @tornado.web.authenticated
-    def get(self):
-
-        search_str = self.get_argument('search_str')
-        print('Searching for', search_str)
-
-        results = fairworkflows.Workflowhub.search(search_str)
-
-        ret = json.dumps(results)
-        self.finish(ret)
-
-
-def workflowhub_search_handler(base_url='/'):
-    endpoint = url_path_join(base_url, '/workflowhub')
-    return endpoint, WorkflowhubSearchHandler
-
 
 class NanopubStepHandler(APIHandler):
